[parser] [fix] Restructure test files
Mirror folder structure of parser/ into tests/parser/ so that each script has a direct mapping to its test file.
