+++ package.json
@@ -4,7 +4,7 @@
   "description": "A scripting language designed for game story writers.",
   "main": "index.js",
   "scripts": {
-    "test": "eslint index.js lib/** tests/** && mocha --compilers js:babel-core/register tests",
+    "test": "eslint index.js lib/** tests/** && mocha --compilers js:babel-core/register tests/**/*.js",
     "fix": "eslint index.js lib/** tests/** --fix"
   },
   "repository": {
